[refactor🔨] : 멤버 연관관계 수정, 회원가입 테스트 수정

+++ src/main/java/shop/gaship/gashipshoppingmall/member/dto/MemberCreationRequest.java
@@ -18,7 +18,7 @@ import lombok.Data;
 @Data
 @Builder
 public class MemberCreationRequest implements MemberDto {
-    private Integer recommendMemberNo;
+    private Long recommendMemberNo;
 
     private String email;
